Allow runtime to bind private math intrinsics
