+++ include/vm/vm.h
@@ -300,8 +300,6 @@ typedef struct {
     TypeKind returnType;
 } IntrinsicSignatureInfo;
 
-const IntrinsicSignatureInfo* vm_get_intrinsic_signature(const char* symbol);
-
 // Function
 typedef struct {
     int start;
@@ -317,6 +315,9 @@ typedef struct {
     int arity;
     Type* returnType;
 } NativeFunction;
+
+const IntrinsicSignatureInfo* vm_get_intrinsic_signature(const char* symbol);
+NativeFn vm_lookup_core_intrinsic(const char* symbol);
 
 // Forward declaration for typed register windows used by call frames.
 typedef struct TypedRegisterWindow TypedRegisterWindow;
@@ -425,6 +426,7 @@ typedef struct {
     Type* type;
     int function_index;
     char* intrinsic_symbol;
+    bool is_internal_intrinsic;
 } ModuleExportEntry;
 
 #define MODULE_EXPORT_NO_REGISTER UINT16_MAX
@@ -996,6 +998,7 @@ static inline void compiler_reset_exports(Compiler* compiler) {
         compiler->exports[i].register_index = -1;
         compiler->exports[i].function_index = -1;
         compiler->exports[i].intrinsic_symbol = NULL;
+        compiler->exports[i].is_internal_intrinsic = false;
         if (compiler->exports[i].type) {
             module_free_export_type(compiler->exports[i].type);
             compiler->exports[i].type = NULL;
