+++ src/compiler/backend/codegen/statements.c
@@ -1196,14 +1196,20 @@ void compile_statement(CompilerContext* ctx, TypedASTNode* stmt) {
             compile_for_iter_statement(ctx, stmt);
             break;
             
-        case NODE_FUNCTION:
-            if (!ctx->compiling_function && stmt->original->function.isPublic &&
-                !stmt->original->function.isMethod && stmt->original->function.name) {
+        case NODE_FUNCTION: {
+            bool is_public = stmt->original->function.isPublic;
+            bool is_core_intrinsic = stmt->original->function.hasCoreIntrinsic;
+            if (!ctx->compiling_function && !stmt->original->function.isMethod &&
+                stmt->original->function.name && (is_public || is_core_intrinsic)) {
                 record_module_export(ctx, stmt->original->function.name, MODULE_EXPORT_KIND_FUNCTION,
                                      stmt->resolvedType);
+                if (is_core_intrinsic && !is_public) {
+                    mark_module_export_internal_intrinsic(ctx, stmt->original->function.name);
+                }
             }
             compile_function_declaration(ctx, stmt);
             break;
+        }
 
         case NODE_IMPORT:
             compile_import_statement(ctx, stmt);
