+++ tests/modules/resolver/README.md
@@ -4,7 +4,8 @@ This directory exercises the module resolution pipeline.
 
 * `default_std_import.orus` verifies that bare module names resolve to the
   bundled `std/` directory next to the executable. It prints boolean checks for
-  `math.add`, `math.scale`, and `math.PI`.
+  the floating-point math wrappers (`math.sin`, `math.cos`, `math.pow`,
+  `math.sqrt`) and constants (`math.PI`, `math.E`).
 * `selective_std_import.orus` ensures selective imports pick individual symbols
   from the same standard module.
 * `oruspath_override.orus` demonstrates overriding the search roots with the
